Get rid of `winapi` from daemon
For this, we needed to import `EXCEPTION_EXECUTE_HANDLER` from
`windows-sys` instead of `winapi`, which is available in `windows-sys >=
0.59`.

Update `socket2` to bring the transitive dep on `windows-sys` to `0.59`

+++ talpid-wireguard/src/wireguard_nt/daita.rs
@@ -394,7 +394,7 @@ impl Machinist {
 
         let wait_events = [
             self.quit_event.as_raw(),
-            self.daita.event_data_available_event() as isize,
+            self.daita.event_data_available_event(),
         ];
 
         let mut event_buffer: [Event; EVENTS_CAPACITY] = unsafe { std::mem::zeroed() };
